fix float bug

+++ src/main/java/onefengma/demo/server/model/product/ShopBrief.java
@@ -3,6 +3,8 @@ package onefengma.demo.server.model.product;
 import onefengma.demo.server.core.LogUtils;
 import onefengma.demo.server.services.funcs.CityDataHelper;
 
+import java.math.BigDecimal;
+
 /**
  * Created by chufengma on 16/6/18.
  */
@@ -13,13 +15,13 @@ public class ShopBrief {
     public String cover;
     public String ironTypeDesc;
     public String handingTypeDesc;
-    public float ironCount;
-    public float handingCount;
-    public float ironMoney;
-    public float handingMoney;
+    public BigDecimal ironCount;
+    public BigDecimal handingCount;
+    public BigDecimal ironMoney;
+    public BigDecimal handingMoney;
 
-    public float money;
-    public float count;
+    public BigDecimal money;
+    public BigDecimal count;
     public float score;
 
     public int productNumbers;
